FABRIC-680: Fixed some of the ignored tests in tooling/tooling-camel-face.

+++ tooling/tooling-camel-facade/src/main/java/org/fusesource/fabric/camel/facade/mbean/CamelThreadPoolMBean.java
@@ -51,7 +51,7 @@ public interface CamelThreadPoolMBean {
 
     boolean isTaskQueueEmpty();
 
-    long getKeepAliveTime();
+    int getKeepAliveTime();
 
     void setKeepAliveTime(int keepAliveTimeInSeconds);
